Move to rpcTests
